+++ src/main/java/com/diggindie/vote/domain/team/controller/TeamController.java
@@ -1,6 +1,7 @@
 package com.diggindie.vote.domain.team.controller;
 
 import com.diggindie.vote.common.code.SuccessCode;
+import com.diggindie.vote.common.config.security.CustomUserDetails;
 import com.diggindie.vote.common.response.Response;
 import com.diggindie.vote.domain.team.dto.TeamListResponse;
 import com.diggindie.vote.domain.team.dto.TeamVoteRequestDto;
@@ -12,7 +13,10 @@ import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.concurrent.TimeUnit;
@@ -41,25 +45,19 @@ public class TeamController {
         return ResponseEntity.ok().body(response);
     }
 
-    public void vote(String externalId, TeamVoteRequestDto request) {
-        String lockKey = TEAM_VOTE_LOCK_PREFIX + externalId; // 투표자 기준 락
-        RLock lock = redissonClient.getLock(lockKey);
-
-        try {
-            boolean acquired = lock.tryLock(3, 15, TimeUnit.SECONDS); // wait 3s, lease 15s
-            if (!acquired) {
-                throw new IllegalStateException("요청이 많습니다. 잠시 후 다시 시도해주세요.");
-            }
-
-            teamVoteExecutor.execute(externalId, request);
-            log.info("팀 투표 완료 - externalId: {}, teamId: {}", externalId, request.teamId());
-
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            throw new IllegalStateException("투표 처리 중 오류가 발생했습니다.");
-        } finally {
-            if (lock.isHeldByCurrentThread()) lock.unlock();
-        }
+    @PreAuthorize("isAuthenticated()")
+    @PostMapping("/votes/teams")
+    public ResponseEntity<Response<Void>> voteTeam(
+            @AuthenticationPrincipal CustomUserDetails userDetails,
+            @RequestBody TeamVoteRequestDto request
+    ) {
+        teamService.vote(userDetails.getUserId(), request);
+        return ResponseEntity.ok().body(Response.of(
+                SuccessCode.INSERT_SUCCESS,
+                true,
+                "팀 투표 완료",
+                (Void) null
+        ));
     }
 }
